Make ppc64le fix also work for x86/amd
The added build tag pulled in ppc64le, but it broke on my
x86_64 system. Build tags are pulled from file names as well,
so let's just use the ones in the file here to simplify.
Otherwise, the file name would have started to get a little out of
control.

Signed-off-by: Christy Perez <christy@linux.vnet.ibm.com>